+++ src/contacts/ContactsContainer.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import ReactRouterPropTypes from 'react-router-prop-types';
 import Switch from 'react-router-dom/Switch';
 import Route from 'react-router-dom/Route';
 import { FormattedMessage } from 'react-intl';
@@ -10,36 +10,41 @@ import {
 } from '@folio/stripes/components';
 
 import ViewContact from './ViewContact';
-import EditContactContainer from './EditContact';
+import EditContact from './EditContact';
 
 class ContactsContainer extends Component {
   static propTypes = {
-    match: PropTypes.object.isRequired,
-    stripes: PropTypes.object.isRequired,
+    match: ReactRouterPropTypes.match.isRequired,
+    history: ReactRouterPropTypes.history.isRequired,
   };
 
   constructor(props, context) {
     super(props, context);
-    this.connectedViewContact = props.stripes.connect(ViewContact);
-    this.connectedEditContactContainer = props.stripes.connect(EditContactContainer);
+
     this.callout = React.createRef();
   }
 
+  onClose = (orgId, contactId) => {
+    if (!contactId) {
+      this.props.history.push(`/organizations/view/${orgId}`);
+    } else {
+      this.props.history.push(`/organizations/${orgId}/contacts/details/${contactId}/view`);
+    }
+  };
+
   goToEdit = (props) => (
-    <this.connectedEditContactContainer
+    <EditContact
       {...props}
       orgId={this.props.match.params.orgId}
       showMessage={this.showMessage}
-      stripes={this.props.stripes}
     />
   );
 
   goToView = (props) => (
-    <this.connectedViewContact
+    <ViewContact
       {...props}
       orgId={this.props.match.params.orgId}
       showMessage={this.showMessage}
-      stripes={this.props.stripes}
       baseUrl={this.props.match.url}
     />
   );
@@ -63,8 +68,31 @@ class ContactsContainer extends Component {
             render={this.goToView}
           />
           <Route
+            path={`${url}/details/:id/view`}
+            render={(props) => (
+              <ViewContact
+                {...props}
+                orgId={this.props.match.params.orgId}
+                showMessage={this.showMessage}
+                baseUrl={`${this.props.match.url}/details`}
+                onClose={this.onClose}
+              />
+            )}
+          />
+          <Route
             path={`${url}/:id/edit`}
             render={this.goToEdit}
+          />
+          <Route
+            path={`${url}/details/:id/edit`}
+            render={(props) => (
+              <EditContact
+                {...props}
+                orgId={this.props.match.params.orgId}
+                showMessage={this.showMessage}
+                onClose={this.onClose}
+              />
+            )}
           />
           <Route
             path={`${url}/add-contact`}