+++ src/contacts/ViewContact/ViewContact.js
@@ -3,21 +3,47 @@ import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
 
 import {
-  Pane,
-  Icon,
+  Accordion,
+  AccordionSet,
   Button,
-  Row,
   Col,
+  ExpandAllButton,
+  Icon,
+  Pane,
+  Row,
 } from '@folio/stripes/components';
 import { AppIcon } from '@folio/stripes/core';
+import { useAccordionToggle } from '@folio/stripes-acq-components';
 
-import ContactPerson from '../../ContactPeople/ContactPerson';
+import {
+  hydrateAddresses,
+  transformCategoryIdsToLables,
+} from '../../common/utils';
+import {
+  CONTACT_PERSON_ACCORDIONS,
+  CONTACT_PERSON_ACCORDION_LABELS,
+} from '../constants';
+import ContactDetails from './ContactDetails';
+import ContactAddresses from './ContactAddresses';
+import ContactEmails from './ContactEmails';
+import ContactUrls from './ContactUrls';
+import ContactPhoneNumbers from './ContactPhoneNumbers';
 
-const ViewContact = ({ onClose, contact, categories, baseUrl, unassign, deleteContact }) => {
+const ViewContact = ({
+  categories,
+  onClose,
+  contact,
+  editUrl,
+  unassign,
+  deleteContact,
+}) => {
+  const [expandAll, sections, toggleSection] = useAccordionToggle();
+  const contactCategories = transformCategoryIdsToLables(categories, contact.categories);
+  const addresses = hydrateAddresses(categories, contact.addresses);
+
   // eslint-disable-next-line react/prop-types
   const getActionMenu = ({ onToggle }) => {
     const contactId = contact.id;
-    const editUrl = `${baseUrl}/${contactId}/edit`;
 
     return (
       <div data-test-view-contact-actions>
@@ -81,13 +107,74 @@ const ViewContact = ({ onClose, contact, categories, baseUrl, unassign, deleteCo
       onClose={onClose}
       paneTitle={`${contact.lastName}, ${contact.firstName}`}
     >
-      <Row>
+      <Row data-test-contact-person>
         <Col xs={12} md={8} mdOffset={2}>
-          <ContactPerson
-            contact={contact}
-            categories={categories}
-            withCollapsing={false}
-          />
+          <Row end="xs">
+            <Col xs={12}>
+              <ExpandAllButton
+                accordionStatus={sections}
+                onToggle={expandAll}
+              />
+            </Col>
+          </Row>
+          <AccordionSet
+            accordionStatus={sections}
+            onToggle={toggleSection}
+          >
+            <Accordion
+              id={CONTACT_PERSON_ACCORDIONS.NAME}
+              label={CONTACT_PERSON_ACCORDION_LABELS[CONTACT_PERSON_ACCORDIONS.NAME]}
+            >
+              <ContactDetails
+                firstName={contact.firstName}
+                lastName={contact.lastName}
+                prefix={contact.prefix}
+                language={contact.language}
+                isInactive={contact.inactive}
+                categories={contactCategories}
+                notes={contact.notes}
+              />
+            </Accordion>
+
+            <Accordion
+              id={CONTACT_PERSON_ACCORDIONS.EMAILS}
+              label={CONTACT_PERSON_ACCORDION_LABELS[CONTACT_PERSON_ACCORDIONS.EMAILS]}
+            >
+              <ContactEmails
+                categories={categories}
+                emails={contact.emails}
+              />
+            </Accordion>
+
+            <Accordion
+              id={CONTACT_PERSON_ACCORDIONS.PHONES}
+              label={CONTACT_PERSON_ACCORDION_LABELS[CONTACT_PERSON_ACCORDIONS.PHONES]}
+            >
+              <ContactPhoneNumbers
+                categories={categories}
+                phoneNumbers={contact.phoneNumbers}
+              />
+            </Accordion>
+
+            <Accordion
+              id={CONTACT_PERSON_ACCORDIONS.URLS}
+              label={CONTACT_PERSON_ACCORDION_LABELS[CONTACT_PERSON_ACCORDIONS.URLS]}
+            >
+              <ContactUrls
+                categories={categories}
+                urls={contact.urls}
+              />
+            </Accordion>
+
+            <Accordion
+              id={CONTACT_PERSON_ACCORDIONS.ADDRESSES}
+              label={CONTACT_PERSON_ACCORDION_LABELS[CONTACT_PERSON_ACCORDIONS.ADDRESSES]}
+            >
+              <ContactAddresses
+                addresses={addresses}
+              />
+            </Accordion>
+          </AccordionSet>
         </Col>
       </Row>
     </Pane>
@@ -97,10 +184,14 @@ const ViewContact = ({ onClose, contact, categories, baseUrl, unassign, deleteCo
 ViewContact.propTypes = {
   onClose: PropTypes.func.isRequired,
   contact: PropTypes.object.isRequired,
-  categories: PropTypes.arrayOf(PropTypes.object),
-  baseUrl: PropTypes.string.isRequired,
+  editUrl: PropTypes.string.isRequired,
   unassign: PropTypes.func.isRequired,
   deleteContact: PropTypes.func.isRequired,
+  categories: PropTypes.arrayOf(PropTypes.object),
+};
+
+ViewContact.defaultProps = {
+  categories: [],
 };
 
 export default ViewContact;