+++ src/contacts/constants.js
@@ -11,8 +11,8 @@ export const CONTACT_PERSON_ACCORDIONS = {
 
 export const CONTACT_PERSON_ACCORDION_LABELS = {
   [CONTACT_PERSON_ACCORDIONS.NAME]: <FormattedMessage id="ui-organizations.contactPeople.name" />,
-  [CONTACT_PERSON_ACCORDIONS.EMAILS]: <FormattedMessage id="ui-organizations.contactPeople.email" />,
-  [CONTACT_PERSON_ACCORDIONS.PHONES]: <FormattedMessage id="ui-organizations.contactPeople.phone" />,
-  [CONTACT_PERSON_ACCORDIONS.URLS]: <FormattedMessage id="ui-organizations.contactPeople.url" />,
-  [CONTACT_PERSON_ACCORDIONS.ADDRESSES]: <FormattedMessage id="ui-organizations.contactPeople.address" />,
+  [CONTACT_PERSON_ACCORDIONS.EMAILS]: <FormattedMessage id="ui-organizations.contactPeople.emails" />,
+  [CONTACT_PERSON_ACCORDIONS.PHONES]: <FormattedMessage id="ui-organizations.contactPeople.phoneNumbers" />,
+  [CONTACT_PERSON_ACCORDIONS.URLS]: <FormattedMessage id="ui-organizations.contactPeople.urls" />,
+  [CONTACT_PERSON_ACCORDIONS.ADDRESSES]: <FormattedMessage id="ui-organizations.contactPeople.addresses" />,
 };